Parsing: split doc comments to separate file
